// frontend/src/types/generated/models/TurningPoint.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { TurningPointType } from "./TurningPointType";

/**
 * 关键转折点
 */
export type TurningPoint = { round: number, eventType: TurningPointType, description: string, passRateBefore: number | null, passRateAfter: number | null, timestamp: string, };